+++ CHANGELOG.md
@@ -27,7 +27,8 @@ Thank you to all who have contributed!
 
 ### Added
 - Added SqlDialect support for MAP expression and type serialization
-
+- Added MAP functions: contains_key, map_get, map_keys, map_values, map_entries, size, exists, cardinality and `IS MAP<K, V>
+ 
 ### Changed
 
 ### Deprecated
@@ -40,6 +41,7 @@ Thank you to all who have contributed!
 
 ### Contributors
 Thank you to all who have contributed!
+- @xd1313113
 
 ## [1.4.0](https://github.com/partiql/partiql-lang-kotlin/releases/tag/v1.4.0) - 2026-06-05